refactor: Flatten core package structure and integrate color utilities directly into the package.
