--- packages/core/colors/CHANGELOG.md
@@ -1,54 +1,0 @@
-# @xaui/colors
-
-## 0.1.5
-
-### Patch Changes
-
-- 2fd4e38: docs: update colors documentation
-
-## 0.1.4
-
-### Patch Changes
-
-- 3011107: Add Button and IconButton components with comprehensive features
-  - Implement Button component with 6 variants (solid, outlined, flat, light, elevated, faded)
-  - Add IconButton component for icon-only actions with default transparent background
-  - Add ripple effect animation using native Animated API
-  - Support loading states with CircularActivityIndicator integration
-  - Include 7 theme colors (primary, secondary, tertiary, danger, warning, success, default)
-  - Provide 4 size options (xs, sm, md, lg) and 5 radius options
-  - Add comprehensive test suite with 17 passing tests
-  - Include detailed documentation with examples and API reference
-  - Update demo app with button component examples
-
-  Add Progress component with comprehensive features
-  - Implement CircularProgress and LinearProgress components
-  - Add support for custom colors and sizes
-  - Include loading states and progress tracking
-  - Add comprehensive test suite with 10 passing tests
-  - Include detailed documentation with examples and API reference
-  - Update demo app with progress component examples
-
-## 0.1.3
-
-### Patch Changes
-
-- a43f64b: Add keywords to improve package discoverability
-
-  Added comprehensive keywords including colors, color-palette, tailwind, tailwindcss, typescript, design-system, ui, react-native, xaui, theming, color-scheme, and design-tokens to help users find the package more easily on npm.
-
-## 0.1.2
-
-### Patch Changes
-
-- d07bd98: Add package description to specify TypeScript colors list utility
-
-  Added description field to package.json to better communicate that @xaui/colors is a TypeScript colors list utility inspired by Tailwind CSS color palette with 20+ color families and 11 shades each (50-950).
-
-## 0.1.1
-
-### Patch Changes
-
-- 126894c: Rename package scope from @xui to @xaui
-
-  The @xui npm scope was already taken, so all packages have been renamed to use the @xaui scope instead. This is a breaking change in terms of package name, but the API remains unchanged.
