--- packages/core/colors/eslint.config.js
@@ -1,3 +1,0 @@
-import baseConfig from '../../../eslint.config.base.js'
-
-export default [...baseConfig]
