--- packages/core/colors/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@xaui/colors",
-  "version": "0.1.5",
-  "description": "A TypeScript colors list utility inspired by Tailwind CSS color palette with 20+ color families and 11 shades each",
-  "keywords": [
-    "colors",
-    "color-palette",
-    "tailwind",
-    "tailwindcss",
-    "typescript",
-    "design-system",
-    "ui",
-    "react-native",
-    "xaui",
-    "theming",
-    "color-scheme",
-    "design-tokens"
-  ],
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist",
-    "package.json"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/rygrams/xaui.git",
-    "directory": "packages/colors"
-  },
-  "scripts": {
-    "build": "tsup src/index.ts --format cjs,esm --dts",
-    "dev": "tsup src/index.ts --format cjs,esm --dts --watch",
-    "test": "vitest",
-    "test:ui": "vitest --ui",
-    "test:coverage": "vitest --coverage",
-    "lint": "eslint src/",
-    "type-check": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "@types/node": "^20.11.17",
-    "tsup": "^8.5.1",
-    "typescript": "^5.9.3"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
