--- packages/core/colors/vitest.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: {
-    globals: true,
-    environment: 'node',
-    include: ['src/**/*.{test,spec}.{js,ts}'],
-  },
-})
