--- packages/core/core/.gitignore
@@ -1,5 +1,0 @@
-dist
-node_modules
-*.log
-.turbo
-coverage
