--- packages/core/core/src/index.ts
@@ -1,21 +1,0 @@
-export { XUIProvider, colors, getColor } from './theme'
-
-export type { ColorName, ColorShade, XUIProviderProps } from './theme'
-
-export { theme, darkTheme } from './theme-config'
-
-export type {
-  XUITheme,
-  ThemeSpacing,
-  ThemeBorderRadius,
-  ThemeBorderWidth,
-  ThemeFontSizes,
-  ThemeFontWeights,
-  ThemeFontFamilies,
-  ThemeShadows,
-} from './theme-config'
-
-export { themeColors, darkThemeColors } from './theme-colors'
-export type { ThemeColors, ColorScheme } from './theme-colors'
-
-export { useXUITheme } from './use-theme'
