--- packages/core/core/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "jsx": "react-native"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist"]
-}
