+++ packages/core/eslint.config.js
@@ -1,0 +1,3 @@
+import baseConfig from '../../eslint.config.base.js'
+
+export default [...baseConfig]
