+++ packages/core/package.json
