+++ packages/core/src/__tests__/__mocks__/react-native.ts
