+++ packages/core/src/theme/use-theme.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import { XUIThemeContext } from './theme'
+import { XUIThemeContext } from './theme-provider'
 import type { XUITheme } from './theme-config'
 
 export function useXUITheme(): XUITheme {
