+++ packages/core/vitest.config.ts
