+++ pnpm-workspace.yaml
@@ -1,6 +1,5 @@
 packages:
-  - packages/components/*
-  - packages/core/*
+  - packages/*
   - apps/*
 
 nodeLinker: hoisted
